requirement.txt
GPU, Python==3.8.16
PyTorch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
matplotlib==3.7.1